Bug 1301477 - Convert artifact model into a set of etl methods
